Split core out into AbstractChainRules, Requires-ize 3rd party packages. Reorganize all files
