--- REQUIRE
@@ -1,4 +1,0 @@
-julia 1.0.0
-SpecialFunctions
-NaNMath
-Cassette 0.2.0 0.3.0
