+++ src/rulesets/Base/mapreduce.jl
