+++ src/rulesets/LinearAlgebra/structured.jl
