+++ test/rulesets/Base/mapreduce.jl
