+++ test/rulesets/LinearAlgebra/structured.jl
